Add HandlerFunction type and improve response handling
Two key updates were made in this commit. Firstly, the HandlerFunction type was added to 'jargonEndpoint.ts' and imported into 'requestHandler.ts', enhancing function definition clarity and type safety. Secondly, improvements were made to the response handling in 'requestHandler.ts'. The status in the response is conditionally set to 204 if the body of the response is null, ensuring appropriate status codes are returned. The headers' typing was also refined to provide better type safety.