Add support for JargonResponse in framework and update /users POST endpoint
In order to provide more flexibility and structure for responses, this commit introduces the "JargonResponse" class in the framework. This handles setting the status code, headers, and body of a response. Following the addition of JargonResponse, the /users POST endpoint is updated to use this new method for constructing responses. This shift towards a defined response structure will streamline response handling and provide a concrete pattern for future development.