Add users table and refactor error handling
This commit accomplishes two main tasks:
1. We have created the "users" DB table and its metadata.
The newly added SQL file creates a "users" table with fields for id, full_name, phone, role, and timestamps. The corresponding snapshot and journal JSON files provide an up-to-date snapshot of our database schema and a log of our schema changes, respectively.

2. Moved http error handling code
The Http error classes have been moved to the framework directory and renamed for better categorization and easier access. Subsequently, all references to these classes are updated to reflect the new location. The aim here is to improve organization and maintainability of code.